requirements.txt
accelerate
codetiming
datasets
# flash-attn>=2.4.3
liger-kernel
mathruler
numpy
omegaconf
pandas
peft
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils
ray[default]
tensordict
torchdata
transformers>=4.51.0,<4.53.0
vllm==0.8.5.post1
wandb
